refactor(controller): Move scheduleOnDifferentHostNodesLabel feature to separate files
