+++ pkg/apis/workflow/v1alpha1/workflow_types.go
@@ -1204,6 +1204,41 @@ func SucceededPodNode(n NodeStatus) bool {
 	return n.Type == NodeTypePod && n.Phase == NodeSucceeded
 }
 
+// Children returns the children of the parent.
+func (s Nodes) Children(parentNodeId string) Nodes {
+	var childNodes = make(Nodes)
+	parentNode, ok := s[parentNodeId]
+	if !ok {
+		return childNodes
+	}
+	for _, childID := range parentNode.Children {
+		if childNode, ok := s[childID]; ok {
+			childNodes[childID] = childNode
+		}
+	}
+	return childNodes
+}
+
+// Filter returns the subset of the nodes that match the predicate, e.g. only failed nodes
+func (s Nodes) Filter(predicate func(NodeStatus) bool) Nodes {
+	var filteredNodes = make(Nodes)
+	for _, node := range s {
+		if predicate(node) {
+			filteredNodes[node.Name] = node
+		}
+	}
+	return filteredNodes
+}
+
+// Map maps the nodes to new values, e.g. `x.Hostname`
+func (s Nodes) Map(f func(x NodeStatus) interface{}) map[string]interface{} {
+	var values = make(map[string]interface{})
+	for _, node := range s {
+		values[node.Name] = f(node)
+	}
+	return values
+}
+
 // UserContainer is a container specified by a user.
 type UserContainer struct {
 	apiv1.Container `json:",inline" protobuf:"bytes,1,opt,name=container"`
