+++ workflow/util/util_test.go
@@ -8,7 +8,6 @@ import (
 	"testing"
 
 	"github.com/stretchr/testify/assert"
-	apiv1 "k8s.io/api/core/v1"
 	v1 "k8s.io/api/core/v1"
 	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
 	"k8s.io/apimachinery/pkg/apis/meta/v1/unstructured"
@@ -473,272 +472,6 @@ func TestGetNodeType(t *testing.T) {
 		assert.NotEqual(t, wfv1.NodeTypePod, GetNodeType(&wfv1.Template{Steps: []wfv1.ParallelSteps{}}))
 		assert.NotEqual(t, wfv1.NodeTypePod, GetNodeType(&wfv1.Template{DAG: &wfv1.DAGTemplate{}}))
 		assert.NotEqual(t, wfv1.NodeTypePod, GetNodeType(&wfv1.Template{Suspend: &wfv1.SuspendTemplate{}}))
-	})
-}
-
-func TestRemoveDuplicates(t *testing.T) {
-	t.Run("EmptySlice", func(t *testing.T) {
-		assert.Equal(t, []string{}, RemoveDuplicates([]string{}))
-	})
-	t.Run("RemoveDuplicates", func(t *testing.T) {
-		assert.Equal(t, []string{"a", "c", "d"}, RemoveDuplicates([]string{"a", "c", "c", "d"}))
-	})
-}
-
-func TestAddHostnamesToAffinity(t *testing.T) {
-	hostNames := []string{"hostnameA", "hostnameB", "hostnameC"}
-	hostSelector := "kubernetes.io/hostname"
-
-	t.Run("EmptyAffinity", func(t *testing.T) {
-		type retryNode struct {
-			targetAffinity *apiv1.Affinity
-		}
-
-		targetNode := &retryNode{}
-		targetNode.targetAffinity = AddHostnamesToAffinity(hostSelector, hostNames, targetNode.targetAffinity)
-		targetNodeSelectorRequirement :=
-			targetNode.targetAffinity.NodeAffinity.RequiredDuringSchedulingIgnoredDuringExecution.NodeSelectorTerms[0].MatchExpressions[0]
-		sourceNodeSelectorRequirement := apiv1.NodeSelectorRequirement{
-			Key:      hostSelector,
-			Operator: apiv1.NodeSelectorOpNotIn,
-			Values:   hostNames,
-		}
-		assert.Equal(t, sourceNodeSelectorRequirement, targetNodeSelectorRequirement)
-	})
-	t.Run("EmptyNodeAffinity", func(t *testing.T) {
-		type retryNode struct {
-			targetAffinity *apiv1.Affinity
-		}
-		targetNode := &retryNode{
-			targetAffinity: &apiv1.Affinity{
-				NodeAffinity: &apiv1.NodeAffinity{},
-			},
-		}
-		targetNode.targetAffinity = AddHostnamesToAffinity(hostSelector, hostNames, targetNode.targetAffinity)
-		targetNodeSelectorRequirement :=
-			targetNode.targetAffinity.NodeAffinity.RequiredDuringSchedulingIgnoredDuringExecution.NodeSelectorTerms[0].MatchExpressions[0]
-		sourceNodeSelectorRequirement := apiv1.NodeSelectorRequirement{
-			Key:      hostSelector,
-			Operator: apiv1.NodeSelectorOpNotIn,
-			Values:   hostNames,
-		}
-		assert.Equal(t, sourceNodeSelectorRequirement, targetNodeSelectorRequirement)
-	})
-	t.Run("EmptyRequiredDuringSchedulingIgnoredDuringExecution", func(t *testing.T) {
-		type retryNode struct {
-			targetAffinity *apiv1.Affinity
-		}
-		targetNode := &retryNode{
-			targetAffinity: &apiv1.Affinity{
-				NodeAffinity: &apiv1.NodeAffinity{
-					RequiredDuringSchedulingIgnoredDuringExecution: &apiv1.NodeSelector{},
-				},
-			},
-		}
-		targetNode.targetAffinity = AddHostnamesToAffinity(hostSelector, hostNames, targetNode.targetAffinity)
-		targetNodeSelectorRequirement :=
-			targetNode.targetAffinity.NodeAffinity.RequiredDuringSchedulingIgnoredDuringExecution.NodeSelectorTerms[0].MatchExpressions[0]
-		sourceNodeSelectorRequirement := apiv1.NodeSelectorRequirement{
-			Key:      hostSelector,
-			Operator: apiv1.NodeSelectorOpNotIn,
-			Values:   hostNames,
-		}
-		assert.Equal(t, sourceNodeSelectorRequirement, targetNodeSelectorRequirement)
-	})
-	t.Run("EmptyNodeSelectorTerms", func(t *testing.T) {
-		type retryNode struct {
-			targetAffinity *apiv1.Affinity
-		}
-		targetNode := &retryNode{
-			targetAffinity: &apiv1.Affinity{
-				NodeAffinity: &apiv1.NodeAffinity{
-					RequiredDuringSchedulingIgnoredDuringExecution: &apiv1.NodeSelector{
-						NodeSelectorTerms: []apiv1.NodeSelectorTerm{},
-					},
-				},
-			},
-		}
-		targetNode.targetAffinity = AddHostnamesToAffinity(hostSelector, hostNames, targetNode.targetAffinity)
-		targetNodeSelectorRequirement :=
-			targetNode.targetAffinity.NodeAffinity.RequiredDuringSchedulingIgnoredDuringExecution.NodeSelectorTerms[0].MatchExpressions[0]
-		sourceNodeSelectorRequirement := apiv1.NodeSelectorRequirement{
-			Key:      hostSelector,
-			Operator: apiv1.NodeSelectorOpNotIn,
-			Values:   hostNames,
-		}
-		assert.Equal(t, sourceNodeSelectorRequirement, targetNodeSelectorRequirement)
-	})
-	t.Run("EmptyMatchExpressions", func(t *testing.T) {
-		type retryNode struct {
-			targetAffinity *apiv1.Affinity
-		}
-		targetNode := &retryNode{
-			targetAffinity: &apiv1.Affinity{
-				NodeAffinity: &apiv1.NodeAffinity{
-					RequiredDuringSchedulingIgnoredDuringExecution: &apiv1.NodeSelector{
-						NodeSelectorTerms: []apiv1.NodeSelectorTerm{
-							{
-								MatchExpressions: []apiv1.NodeSelectorRequirement{},
-							},
-						},
-					},
-				},
-			},
-		}
-		targetNode.targetAffinity = AddHostnamesToAffinity(hostSelector, hostNames, targetNode.targetAffinity)
-		targetNodeSelectorRequirement :=
-			targetNode.targetAffinity.NodeAffinity.RequiredDuringSchedulingIgnoredDuringExecution.NodeSelectorTerms[0].MatchExpressions[0]
-		sourceNodeSelectorRequirement := apiv1.NodeSelectorRequirement{
-			Key:      hostSelector,
-			Operator: apiv1.NodeSelectorOpNotIn,
-			Values:   hostNames,
-		}
-		assert.Equal(t, sourceNodeSelectorRequirement, targetNodeSelectorRequirement)
-	})
-	t.Run("AddMatchExpressions", func(t *testing.T) {
-		type retryNode struct {
-			targetAffinity *apiv1.Affinity
-		}
-		sourceNodeSelectorRequirement0 := apiv1.NodeSelectorRequirement{
-			Key:      "metadata.name",
-			Operator: apiv1.NodeSelectorOpIn,
-			Values:   []string{"hostname1", "hostname2"},
-		}
-		sourceNodeSelectorRequirement1 := apiv1.NodeSelectorRequirement{
-			Key:      hostSelector,
-			Operator: apiv1.NodeSelectorOpNotIn,
-			Values:   hostNames,
-		}
-		targetNode := &retryNode{
-			targetAffinity: &apiv1.Affinity{
-				NodeAffinity: &apiv1.NodeAffinity{
-					RequiredDuringSchedulingIgnoredDuringExecution: &apiv1.NodeSelector{
-						NodeSelectorTerms: []apiv1.NodeSelectorTerm{
-							{
-								MatchExpressions: []apiv1.NodeSelectorRequirement{
-									sourceNodeSelectorRequirement0,
-								},
-							},
-						},
-					},
-				},
-			},
-		}
-		targetNode.targetAffinity = AddHostnamesToAffinity(hostSelector, hostNames, targetNode.targetAffinity)
-		targetNodeSelectorRequirement0 :=
-			targetNode.targetAffinity.NodeAffinity.RequiredDuringSchedulingIgnoredDuringExecution.NodeSelectorTerms[0].MatchExpressions[0]
-		targetNodeSelectorRequirement1 :=
-			targetNode.targetAffinity.NodeAffinity.RequiredDuringSchedulingIgnoredDuringExecution.NodeSelectorTerms[0].MatchExpressions[1]
-		assert.Equal(t, sourceNodeSelectorRequirement0, targetNodeSelectorRequirement0)
-		assert.Equal(t, sourceNodeSelectorRequirement1, targetNodeSelectorRequirement1)
-	})
-	t.Run("AddMatchExpressionsSameHostSelector", func(t *testing.T) {
-		type retryNode struct {
-			targetAffinity *apiv1.Affinity
-		}
-		sourceNodeSelectorRequirement0 := apiv1.NodeSelectorRequirement{
-			Key:      hostSelector,
-			Operator: apiv1.NodeSelectorOpIn,
-			Values:   []string{"hostname1", "hostname2"},
-		}
-		sourceNodeSelectorRequirement1 := apiv1.NodeSelectorRequirement{
-			Key:      hostSelector,
-			Operator: apiv1.NodeSelectorOpNotIn,
-			Values:   hostNames,
-		}
-		targetNode := &retryNode{
-			targetAffinity: &apiv1.Affinity{
-				NodeAffinity: &apiv1.NodeAffinity{
-					RequiredDuringSchedulingIgnoredDuringExecution: &apiv1.NodeSelector{
-						NodeSelectorTerms: []apiv1.NodeSelectorTerm{
-							{
-								MatchExpressions: []apiv1.NodeSelectorRequirement{
-									sourceNodeSelectorRequirement0,
-								},
-							},
-						},
-					},
-				},
-			},
-		}
-		targetNode.targetAffinity = AddHostnamesToAffinity(hostSelector, hostNames, targetNode.targetAffinity)
-		targetNodeSelectorRequirement0 :=
-			targetNode.targetAffinity.NodeAffinity.RequiredDuringSchedulingIgnoredDuringExecution.NodeSelectorTerms[0].MatchExpressions[0]
-		targetNodeSelectorRequirement1 :=
-			targetNode.targetAffinity.NodeAffinity.RequiredDuringSchedulingIgnoredDuringExecution.NodeSelectorTerms[0].MatchExpressions[1]
-		assert.Equal(t, sourceNodeSelectorRequirement0, targetNodeSelectorRequirement0)
-		assert.Equal(t, sourceNodeSelectorRequirement1, targetNodeSelectorRequirement1)
-	})
-	t.Run("AddMatchExpressionsSameOperator", func(t *testing.T) {
-		type retryNode struct {
-			targetAffinity *apiv1.Affinity
-		}
-		sourceNodeSelectorRequirement0 := apiv1.NodeSelectorRequirement{
-			Key:      "metadata.name",
-			Operator: apiv1.NodeSelectorOpNotIn,
-			Values:   []string{"hostname1", "hostname2"},
-		}
-		sourceNodeSelectorRequirement1 := apiv1.NodeSelectorRequirement{
-			Key:      hostSelector,
-			Operator: apiv1.NodeSelectorOpNotIn,
-			Values:   hostNames,
-		}
-		targetNode := &retryNode{
-			targetAffinity: &apiv1.Affinity{
-				NodeAffinity: &apiv1.NodeAffinity{
-					RequiredDuringSchedulingIgnoredDuringExecution: &apiv1.NodeSelector{
-						NodeSelectorTerms: []apiv1.NodeSelectorTerm{
-							{
-								MatchExpressions: []apiv1.NodeSelectorRequirement{
-									sourceNodeSelectorRequirement0,
-								},
-							},
-						},
-					},
-				},
-			},
-		}
-		targetNode.targetAffinity = AddHostnamesToAffinity(hostSelector, hostNames, targetNode.targetAffinity)
-		targetNodeSelectorRequirement0 :=
-			targetNode.targetAffinity.NodeAffinity.RequiredDuringSchedulingIgnoredDuringExecution.NodeSelectorTerms[0].MatchExpressions[0]
-		targetNodeSelectorRequirement1 :=
-			targetNode.targetAffinity.NodeAffinity.RequiredDuringSchedulingIgnoredDuringExecution.NodeSelectorTerms[0].MatchExpressions[1]
-		assert.Equal(t, sourceNodeSelectorRequirement0, targetNodeSelectorRequirement0)
-		assert.Equal(t, sourceNodeSelectorRequirement1, targetNodeSelectorRequirement1)
-	})
-	t.Run("MergeMatchExpressionsWithDuplicates", func(t *testing.T) {
-		type retryNode struct {
-			targetAffinity *apiv1.Affinity
-		}
-		sourceNodeSelectorRequirement0 := apiv1.NodeSelectorRequirement{
-			Key:      hostSelector,
-			Operator: apiv1.NodeSelectorOpNotIn,
-			Values:   []string{"hostname1", "hostname2", "hostnameA"},
-		}
-		sourceNodeSelectorRequirementMerged := apiv1.NodeSelectorRequirement{
-			Key:      hostSelector,
-			Operator: apiv1.NodeSelectorOpNotIn,
-			Values:   []string{"hostname1", "hostname2", "hostnameA", "hostnameB", "hostnameC"},
-		}
-		targetNode := &retryNode{
-			targetAffinity: &apiv1.Affinity{
-				NodeAffinity: &apiv1.NodeAffinity{
-					RequiredDuringSchedulingIgnoredDuringExecution: &apiv1.NodeSelector{
-						NodeSelectorTerms: []apiv1.NodeSelectorTerm{
-							{
-								MatchExpressions: []apiv1.NodeSelectorRequirement{
-									sourceNodeSelectorRequirement0,
-								},
-							},
-						},
-					},
-				},
-			},
-		}
-		targetNode.targetAffinity = AddHostnamesToAffinity(hostSelector, hostNames, targetNode.targetAffinity)
-		targetNodeSelectorRequirement :=
-			targetNode.targetAffinity.NodeAffinity.RequiredDuringSchedulingIgnoredDuringExecution.NodeSelectorTerms[0].MatchExpressions[0]
-		assert.Equal(t, sourceNodeSelectorRequirementMerged, targetNodeSelectorRequirement)
 	})
 }
